% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddBooleanGate.R
\name{addBooleanGate}
\alias{addBooleanGate}
\title{Adds boolean gate to the GatingSet or GatingSetList}
\usage{
addBooleanGate(gs, booleanSubset, parentGate, overrideGate = FALSE,
  booleanGateName = NULL)
}
\arguments{
\item{gs}{Either a GatingSet, a GatingSetList, or path to one of these objects on disk}

\item{booleanSubset}{The booleanSubset (a combination of existing gates) in string format, e.g. "8+/GMM+&!8+/GAMMADELTA"}

\item{parentGate}{The gate under which the booleansubset is added}

\item{booleanGateName}{optional. What to call the new gate}
}
\description{
Adds boolean gate to the GatingSet or GatingSetList
}
