% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxplotSubsetProportionStats.R
\name{boxplot.subset.proportion.stats}
\alias{boxplot.subset.proportion.stats}
\title{Stratified Boxplots of the Proportion of the Boolean Subset seen}
\usage{
\method{boxplot}{subset.proportion.stats}(gsOrGsListOrPath, booleanSubset,
  parentGate, parentGateForProportionCalc = parentGate, groupBy = NULL,
  outdir = NULL, overrideGate = FALSE, ylimits = NULL)
}
\arguments{
\item{gsOrGsListOrPath}{Either a GatingSet, a GatingSetList, or path to one of these objects on disk}

\item{booleanSubset}{The booleanSubset (a combination of existing gates) in string format, e.g. "8+/GMM+&!8+/GAMMADELTA"}

\item{parentGate}{The gate under which the booleansubset is added}

\item{parentGateForProportionCalc}{The parent gate which is to be used in proportion calculations (e.g. "8+" or "3+"). Defaults to parentGate}

\item{groupBy}{optional string. What variable to stratify the boxplot by}

\item{outdir}{Where to save the boxplot and stats, if desired}

\item{overrideGate}{If this is set to TRUE, booleanSubset will override any previous node of the same name under the same parentGate.}

\item{ylimits}{optional numeric vector}
}
\value{
boxplot and stats, unless outdir is specified
}
\description{
Adds the given booleanSubset to the gating tree and creates a stratified boxplot of the per-patient proportions of this subset.
}
\examples{
\dontrun{
boxplot.subset.proportion.stats(gsOrGsListOrPath="/Path/To/GatingSet",
                                booleanSubset="GMM&GAMMADELTA",
                                parentGate="CD3",
                                groupBy="HIVStatus",
                                outdir="Path/To/Plot/Directory",
                                overrideGate=FALSE,
                                ylimits=NULL)
}
}
\keyword{BooleanSubset}
\keyword{Box}
\keyword{Plot}
\keyword{Proportion}
