% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/densityplot.R
\docType{methods}
\name{densityplot,GatingHierarchy,character-method}
\alias{densityplot,GatingHierarchy,character-method}
\title{plot the population as densityplot for the channels associated with the gate}
\usage{
\S4method{densityplot}{GatingHierarchy,character}(x, data, ...)
}
\arguments{
\item{x}{GatingHierarchy}

\item{data}{node/population name}
}
\description{
It is used to check the 1d density of the parent for the purpose of choosing appropriate
1d gating algorithm or fine tune the gating parameters.
}

