% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citrus_helper_functions.R
\name{extract_citrus_clusters}
\alias{extract_citrus_clusters}
\title{This function extracts clusted flow cytometric expression data from an RDS containing 
clusters generated by CITRUS (Bruggner et al, PNAS 2014) and returns them as a named list of matrices
There are two ways to use it:}
\usage{
extract_citrus_clusters(RDSfile, anno, cluster_id = NULL, markers = markers)
}
\arguments{
\item{RDSfile}{path to the Citrus .rds file that contains the Citrus run (eg. "/path/to/Age.rds")}

\item{anno}{data.frame containing annotation. anno <- read.csv("annotation.txt") in directory of each Citrus run}

\item{cluster_id}{numeric id of the cluster to extract , if NULL then extract all cv.min clusters found in RDS}

\item{markers:}{named list translating markers <-> channels}
}
\value{
a named list of expression matrices (list has length 1 if you specified a single cluster, option 2 above)
}
\description{
1. to extract *all* CV.min clusters from the RDSfile, leave cluster_id=NULL
mat_of_cvmin_clusters <- extract_citrus_clusters(""citrusOutput_10000cells/Age.rds", anno, markers)
}
\details{
2. to extract a particular cluster from the RDSfile, specify its cluster_id
mat_of_cluster12345 <- extract_citrus_clusters(""citrusOutput_10000cells/Age.rds", anno, cluster_id=12345, markers)
}

