% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flowWorkspace-methods.R
\name{fixChannelNames}
\alias{fixChannelNames}
\title{post process gs to fix channel name discrepancy caused by letter case#'
1.between gates defined in xml and fcs
2.gates with xml}
\usage{
fixChannelNames(gs)
}
\description{
The assumption is the channel names is consistent across samples in FCS files.
So basically it loop through all gates and match up the gate parameters to
the flow data and update it when needed.
}

