% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flowIncubator.R
\docType{package}
\name{flowIncubator}
\alias{flowIncubator}
\alias{flowIncubator-package}
\title{Routines that don't belong to the core flow packages yet}
\description{
Routines that don't belong to the core flow packages yet.
}
\details{
This package is designed to facilitate other flow core packages.

\tabular{ll}{
Package: \tab flowIncubator\cr
Type: \tab Package\cr
Version: \tab 0.99.6\cr
Date: \tab 2013-08-10\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Mike Jiang \email{wjiang2@fhcrc.org},

Maintainer: Mike Jiang \email{wjiang2@fhcrc.org}
}
\keyword{package}

