% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.gatingML.cytobank.R
\docType{methods}
\name{gating,graphGatingML,GatingSet-method}
\alias{gating,graphGatingML,GatingSet-method}
\title{Apply the gatingML graph to a GatingSet
It applies the gates to the GatingSet based on the population tree described in graphGatingML.}
\usage{
\S4method{gating}{graphGatingML,GatingSet}(x, y, ...)
}
\arguments{
\item{x}{graphGatingML}

\item{y}{GatingSet}

\item{...}{other arguments}
}
\description{
Apply the gatingML graph to a GatingSet
It applies the gates to the GatingSet based on the population tree described in graphGatingML.
}

