% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphGatingML-methods.R
\docType{methods}
\name{gating,graphGML,GatingSet-method}
\alias{gating,graphGML,GatingSet-method}
\title{Apply the gatingML graph to a GatingSet
It applies the gates to the GatingSet based on the population tree described in graphGML.}
\usage{
\S4method{gating}{graphGML,GatingSet}(x, y, ...)
}
\arguments{
\item{x}{graphGML}

\item{y}{GatingSet}

\item{...}{other arguments}
}
\description{
Apply the gatingML graph to a GatingSet
It applies the gates to the GatingSet based on the population tree described in graphGML.
}

