% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingSet2GatingML_hack_version.R
\name{gatingML.compensation}
\alias{gatingML.compensation}
\title{Create a gatingML node for compensation}
\usage{
\method{gatingML}{compensation}(x, y)
}
\description{
Create a gatingML node for compensation
}

