% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GatingSet2GatingML_hack_version.R
\name{gatingML}
\alias{gatingML}
\title{this function retrieves the gates from GatingSet
and writes a customed GatingML-2.0 file
which can then be imported into cytobank
operation:
1. Read in gate geometry, compensation and transformation from gatingSet
2. Rescale gate boundaries with flowJoTrans() so gates show up in flowJo
3. Save gates and hierarchy structure to R environment
4. Write environment out to gatingML using write.GatingML()}
\usage{
gatingML(x, y, ...)
}
\description{
this function retrieves the gates from GatingSet
and writes a customed GatingML-2.0 file
which can then be imported into cytobank
operation:
1. Read in gate geometry, compensation and transformation from gatingSet
2. Rescale gate boundaries with flowJoTrans() so gates show up in flowJo
3. Save gates and hierarchy structure to R environment
4. Write environment out to gatingML using write.GatingML()
}
\examples{
\dontrun{
require(flowWorkspace)
require(XML)
localPath <- "~/rglab/workspace/openCyto"
gs <- load_gs(file.path(localPath,"misc/testSuite/gs-tcell_asinhtGm2"))
outFile <- tempfile()
gatingMLNode <- gatingML(gs)
saveXML(gatingMLNode$value(), encoding = "UTF-8", file = outFile)
}
}

