% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{get_groups}
\alias{get_groups}
\title{Step 1: determine which group to parse for each xml workspace}
\usage{
get_groups(files, pattern)
}
\arguments{
\item{files}{character vector, full paths of `xml` workspace files}

\item{pattern}{character a regular expression to specifiy which group to parse}
}
\value{
a list of opened workspace objects and the group names matched by regular expression.
}
\description{
The regular expression may not always pick up the correct group for each xml.
Thus it is safer to separate this step from parsing so that users get to review and edit(if necessary) 
the resolved group results.
}
\examples{
\dontrun{
 wsFiles <- list.files(dataPath, pattern = ".xml", full = T)
 groups <- get_groups(wsFiles, "(_MD)|(_SK)")
}
}
