% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.gatingML.cytobank.R
\docType{methods}
\name{getCompensationMatrices,graphGatingML-method}
\alias{getCompensationMatrices,graphGatingML-method}
\title{Extract compensation from graphGatingML object.}
\usage{
\S4method{getCompensationMatrices}{graphGatingML}(x)
}
\arguments{
\item{x}{graphGatingML}
}
\value{
compensation object
}
\description{
Extract compensation from graphGatingML object.
}

