% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowWorkspace-methods.R
\name{getDescendants}
\alias{getDescendants}
\title{Convenient wrapper to extract all the descedenat nodes of the given population}
\usage{
getDescendants(gh, node)
}
\arguments{
\item{gh}{GatingHierarchy object}

\item{node}{character that gives the name or path of population node}
}
\description{
Convenient wrapper to extract all the descedenat nodes of the given population
}

