% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getKeywords.R
\docType{methods}
\name{getKeywords,GatingSet,character-method}
\alias{getKeywords,GatingSet,character-method}
\title{Fetch a list of keywords from a GatingSet
Return them as a data.table with columns names for the keyword
The FIL keyword is renamed to 'name' for sample name consistency
No error checking at the moment.}
\usage{
\S4method{getKeywords}{GatingSet,character}(obj, y)
}
\description{
Fetch a list of keywords from a GatingSet
Return them as a data.table with columns names for the keyword
The FIL keyword is renamed to 'name' for sample name consistency
No error checking at the moment.
}

