% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.gatingML.cytobank.R
\docType{methods}
\name{getTransformations,graphGatingML-method}
\alias{getTransformations,graphGatingML-method}
\title{Extract transformations from graphGatingML object.}
\usage{
\S4method{getTransformations}{graphGatingML}(x)
}
\arguments{
\item{x}{graphGatingML}
}
\value{
transformList object
}
\description{
Extract transformations from graphGatingML object.
}

