% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.gatingML.cytobank.R
\docType{class}
\name{graphGatingML-class}
\alias{graphGatingML-class}
\title{It is essentially a graphNEL class and exists for the purpose of method dispatching.}
\description{
It is essentially a graphNEL class and exists for the purpose of method dispatching.
}

