% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flowWorkspace-methods.R
\name{grep.FCS}
\alias{grep.FCS}
\title{a uility function to match fcs files based on the channels used by one example FCS file}
\usage{
grep.FCS(pattern, x)
}
\arguments{
\item{pattern}{\code{character} the example FCS file that contains the channels of interest}

\item{x}{\code{character} vector giving the list of fcs files to match}
}
\value{
a \code{character} vector of fcs files that has the identical channels with \code{subset}
}
\description{
It uses \link{readFCSPar} to read parameters from FCS header to select target files,
thus be used as a prefilter before \code{read.flowSet} or \link{read.ncdfFlowSet} call.
}
\examples{
\dontrun{
grep.FCS(pattern = bcells[2],  x = c(bcells,tcells))
 #return TRUE  TRUE FALSE FALSE
}
}

