% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksstat.R
\name{ksstat}
\alias{ksstat}
\title{kstest
An Rcpp implementation returning the ks statistic between two numeric vectors.}
\usage{
ksstat(x, y)
}
\arguments{
\item{x}{vector a}

\item{y}{vector b}
}
\description{
kstest
An Rcpp implementation returning the ks statistic between two numeric vectors.
}
