% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestSamples.R
\name{nearestSamples}
\alias{nearestSamples}
\title{impute the gate (flagged as failure by external algorithm) with refGate from nearest neighbour sample}
\usage{
nearestSamples(gs, node, failed, passed = NULL, ...)
}
\arguments{
\item{gs}{a \code{GatingSet}}

\item{node}{a \code{character} or \code{numeric} specifing node index}

\item{failed}{a \code{character} or \code{numeric} specifing samples that fails the gating QA}

\item{passed}{a \code{character} or \code{numeric} specifing samples that passes QA to be served as references
By default, all samples other than failed will be used.
but sometime it is helpful to narrow it down to a few of really good samples.}

\item{...}{other arguments passed to \code{.nearestSample}}
}
\description{
impute the gate (flagged as failure by external algorithm) with refGate from nearest neighbour sample
}
\examples{
\dontrun{
library(flowWorkspace)
gs <- load_gs("gs-tcell/")

#Finding reference sample for: 1349_3_Tcell_A06.fcs
res <- nearestSamples(gs, node = "CD3", failed = "1349_3_Tcell_A06.fcs")
# res is a named vector stores the mapping between bad(failed) and good(reference) samples

#fix the gate for the bad samples
regateNearestSamples(gs, res, "CD3")
}
}

