% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{parse_xmls}
\alias{parse_xmls}
\title{Step 2: parse the entire batches}
\usage{
parse_xmls(resolved_groups, output.dir, ...)
}
\arguments{
\item{resolved_groups}{the results returned by 1st step}

\item{...}{other arguments passed to 'parseWorkspace'}
}
\value{
nothing
}
\description{
Sub-steps:
1. parse workspaces into GatingSets 
2. save the parsed GatingSets into "parsed" subfolder under output.dir 
3. report the possible discrepancy within each GatingSet
}
\examples{
\dontrun{
 
 
}
}
