% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowWorkspace-methods.R
\name{plotGate_labkey}
\alias{plotGate_labkey}
\title{plot by prarent index}
\usage{
plotGate_labkey(G, parentID, x, y, smooth = FALSE, cond = NULL,
  xlab = NULL, ylab = NULL, overlay = NULL, overlay.symbol = NULL,
  key = NULL, ...)
}
\arguments{
\item{x}{\code{character} x channel}

\item{y}{\code{character} y channel, if \code{NULL},then try to do \code{densityplot}}
}
\description{
This API is mainly used for labkey module. It takes a parent index instead of the actual gate index.
When there is no gate associated with the x,y channel specified by user, it simply plots the \code{xyplot} 
or \code{densityplot} without the gate.
}

