% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.gatingML.cytobank.R
\name{plotTree}
\alias{plotTree}
\title{plot the population tree}
\usage{
plotTree(g, label = c("popName", "gateName"))
}
\arguments{
\item{g}{a graphNEL generated by constructTree function}

\item{label}{specifies what to be dispaled as node label. Can be either 'popName' (population name parsed from GateSets) or 'gateName'(the name of the actual gate associated with each node)}
}
\description{
plot the population tree
}

