% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestSamples.R
\name{regateNearestSamples}
\alias{regateNearestSamples}
\title{perform the actual regating of identified bad samples with gates from good samples
 using the named vector from flowIncubator:::.nearestSamples()}
\usage{
regateNearestSamples(gs, samples, gate_name)
}
\arguments{
\item{gs}{a \code{GatingSet}}

\item{samples}{a named \code{character} vector of sample names in the supplied GatingSet}

\item{gate_name}{a \code{character} specifying the gate to operate on (eg. "CD3")}
}
\description{
perform the actual regating of identified bad samples with gates from good samples
 using the named vector from flowIncubator:::.nearestSamples()
}
