% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTSNE.R
\name{runOneSense}
\alias{runOneSense}
\title{A function to run One-SENSE. Has an additional dimensionMarkers parameter. For now, assuming people won't run more than 3 dimensions.}
\usage{
runOneSense(gs = NULL, parentGate = NULL, degreeFilterGates = c(),
  otherGates = c(), tsneMarkers = c(), groupBy = c(), degreeFilter = 0,
  seed = 999, theta = 0.9, cloneGs = TRUE, dimensionMarkers = list(),
  ...)
}
\arguments{
\item{dimensionMarkers}{a list of character vectors, containing the markers which are to be run for each One-SENSE dimension}
}
\description{
Note lack of numThreads argument. Multicore tsne doesn't work for One-SENSE (i.e. dims = 1). Since the input matrix dimensions will be smaller, it seems
that multicore tsne might not provide much of a speed-up anyway. See: https://github.com/DmitryUlyanov/Multicore-TSNE#what-to-expect
}
