% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTSNE.R
\name{sampleGatingSetForTsne}
\alias{sampleGatingSetForTsne}
\title{Sample a GatingSet}
\usage{
sampleGatingSetForTsne(gs = NULL, parentGate = NULL, groupBy = c(),
  seed = 999, cloneGs = TRUE)
}
\arguments{
\item{gs}{a GatingSet object or path to a GatingSet directory, properly gated data with annotation in its pData}

\item{parentGate}{a \code{string} describing the gate upstream of the cytokine gates (eg. "CD4", "cd8+", etc...)}

\item{groupBy}{a \code{vector} of \code{strings} describing columns of the \code{gatingSet}'s phenoData.}
}
\value{
a sampled \code{GatingSet}
}
\description{
Sample a GatingSet
}
