% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flowWorkspace-methods.R
\name{save_gslist_labkey}
\alias{save_gslist_labkey}
\title{a wrapper for \code{save_gslist}}
\usage{
save_gslist_labkey(gslist, path, cdf, ...)
}
\value{
a copy of original gslist with modified cdf path when cdf == "move"
otherwise, it behaves the same as \code{save_gslist}
}
\description{
a wrapper for \code{save_gslist}
}

