% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowWorkspace-methods.R
\name{swapChannelMarker}
\alias{swapChannelMarker}
\title{A wrapper that swaps the channel names with marker names}
\usage{
swapChannelMarker(gs)
}
\arguments{
\item{gs}{GatingSet objects}
}
\value{
a new GatingSet
}
\description{
It is useful to prepare multiple GatingSets for merging when the marker (instead of channel) names are consistent accross batches.
It basically updates the dimension names of the underling flow data as well as the gate definitions stored in GatingSets. 
It's important to be aware that the gates in the original GatingSet will be updated even a new GatingSet object will be returend as the result.
}

