% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowWorkspace-methods.R
\name{swapChannelMarker_flowframe}
\alias{swapChannelMarker_flowframe}
\title{Preprocesses a Cytotrol flowFrame object}
\usage{
swapChannelMarker_flowframe(fr, use.exprs = TRUE)
}
\arguments{
\item{fr}{the \code{flowFrame} object to preprocess}

\item{use.exprs}{logical indicate that if the data matrix will be updated as well.}
}
\value{
the updated \code{flowFrame} object containing only the markers of
interest
}
\description{
Our goal here is to use swap the marker names and the channel names within a
\code{flowFrame} object to ensure that the \code{flowFrame} objects across
centers can be merged into a single \code{flowSet}.
}

