% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flowWorkspace-methods.R
\docType{methods}
\name{updateIndices,GatingHierarchy,character,logical-method}
\alias{updateIndices,GatingHierarchy,character,logical-method}
\title{directly update node indices without changing gates}
\usage{
\S4method{updateIndices}{GatingHierarchy,character,logical}(obj, y, z)
}
\arguments{
\item{y}{\code{character} node name or path}

\item{z}{\code{logical} vector as local event indices relative to node \code{y}}

\item{x}{\code{GatingHierarchy} object}
}
\description{
It is useful when we want to alter the popluation at events level yet
without removing or adding the existing gates.
}

