% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxplotSubsetProportionStats.R
\name{boxplot.subset.proportion.stats}
\alias{boxplot.subset.proportion.stats}
\title{Stratified Boxplots of the Proportion of the Boolean/Existing Subset seen}
\usage{
\method{boxplot}{subset.proportion.stats}(gsOrGsListOrPath,
  existingSubset = NULL, booleanSubset = NULL, parentGate,
  parentGateForProportionCalc = parentGate, groupBy = NULL, outdir = NULL,
  overrideGate = FALSE, ylimits = NULL, baseSize = 19,
  sampleIDCol = NULL, jitterWidth = 0.15)
}
\arguments{
\item{gsOrGsListOrPath}{Either a GatingSet, a GatingSetList, or path to one of these objects on disk}

\item{existingSubset}{(existingSubset or booleanSubset must be provided) The name of the existing subset of interest, e.g. "CD3/GMM"}

\item{booleanSubset}{(existingSubset or booleanSubset must be provided) The booleanSubset (a combination of existing gates) in string format, e.g. "8+/GMM+&!8+/GAMMADELTA"}

\item{parentGate}{The gate under which the booleansubset is added}

\item{parentGateForProportionCalc}{The parent gate which is to be used in proportion calculations (e.g. "8+" or "3+"). Defaults to parentGate}

\item{groupBy}{optional string. What variable to stratify the boxplot by}

\item{outdir}{Where to save the boxplot and stats, if desired.}

\item{overrideGate}{If this is set to TRUE, booleanSubset will override any previous node of the same name under the same parentGate.}

\item{ylimits}{optional numeric vector}

\item{sampleIDCol}{(optional) Name of the GatingSet metadata column which contains individual sample identifiers. If sampleIDCol is provided, outliers will be labeled with this value.}

\item{jitterWidth}{}
}
\value{
boxplot and stats, and optionally saves the plot to outdir
}
\description{
Creates a stratified boxplot of the per-patient proportions of the given subset.
Iff booleanSubset is provided, adds the given booleanSubset to the gating tree first.
}
\examples{
\dontrun{
boxplot.subset.proportion.stats(gsOrGsListOrPath="/Path/To/GatingSet",
                                booleanSubset="GMM&GAMMADELTA",
                                parentGate="CD3",
                                groupBy="HIVStatus",
                                outdir="Path/To/Plot/Directory",
                                overrideGate=FALSE,
                                ylimits=NULL)
}
}
\keyword{BooleanSubset}
\keyword{Box}
\keyword{Plot}
\keyword{Proportion}
