% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowWorkspace-methods.R, R/pipeline.R
\name{merge_gs}
\alias{merge_gs}
\alias{merge_gs}
\title{this is the wrapper for labkey where only one gslist to be returned
cluster/merge GatingSets based on the gating tree structures.}
\usage{
merge_gs(output.dir)

merge_gs(output.dir)
}
\arguments{
\item{x}{A \code{list} of \code{GatingSet}s .}
}
\value{
A \code{\link{GatingSetList}} that contains multiple GatingSets each of which share the same gating and data structure.
}
\description{
merge GatingSets based on the gating tree structures.

Sub-steps:
1. load the GatingSets parsed from step.
2 and does validity check to see if the gating trees within each gatingSet is clean
      try to resolve the discrepancy of gating trees within single GatingSet by dropping the extra terminal node for some samples (mistakenly added by the person who does flowJo)
     when there is non-leaf node causing the discrepancy, we throw the error to ask for human intervention such as: correct the typo of population name or drop the entire subtree
3. drop the redundant channels (since they are usually one of the causes for channel discrepancy across batches)
4. further drop the non-leaf nodes that cause the cross-batch discrepancy  
load parsed gs
}
\details{
Group the individual GatingSets by their gating schemes.It is done by comparing the node list returned by \code{\link{getNodes}},which assumes
they follow the same population naming conventions.
Meanwhile the unused channels are automatically dropped to make sure the flow data has identical data structure within each group.
In order to further merge multiple GatingSet objects into one, use \code{\link{rbind2}}.
}
\examples{
\dontrun{
	#load gatingsets from disk
	#gs_toMerge is the path that stores multiple archived gatingsets
	gs_list<-lapply(list.files("flowIncubator/output/gs_toMerge",full=T),function(this_folder){
      flowWorkspace:::load_gs(this_folder)
    })
    
	gs_list <- merge_gs(gs_list)
	gs_list
	
}
}
\seealso{
\code{\link{rbind2}},\code{\link{GatingSetList}}
}
\author{
Mike Jiang \email{wjiang2@fhcrc.org}
}
