% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.gatingML.cytobank.R
\docType{methods}
\name{plot,graphGatingML,missing-method}
\alias{plot,graphGatingML,missing-method}
\title{plot the population tree}
\usage{
\S4method{plot}{graphGatingML,missing}(x, label = c("popName", "gateName"))
}
\arguments{
\item{x}{a graphNEL generated by constructTree function}

\item{label}{specifies what to be dispaled as node label. Can be either 'popName' (population name parsed from GateSets) or 'gateName'(the name of the actual gate associated with each node)}

\item{y}{not used}
}
\description{
plot the population tree
}

