% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flowWorkspace-methods.R
\name{updateGateParameter}
\alias{updateGateParameter}
\title{update the gate parameters for a GatingSet}
\usage{
updateGateParameter(gs, map, nodes = NULL)
}
\arguments{
\item{gs}{\code{GatingSet} to work with}

\item{map}{\code{data.frame} contains the mapping between old and new channel names}

\item{nodes}{\code{character} vector to specify the nodes to be updated. Default is all the nodes}
}
\description{
It actually reconstructs all the gates with their parameters changed
based on given mapping between the old and new channel names.
TODO: we want to have API to only update gate parameters instead of entire gate object.
}

